Move getXataClient to Base
