+++ cli/src/commands/workspaces/list.ts
@@ -1,5 +1,4 @@
 import { Base } from '../../base.js';
-import { getXataClient } from '../../client.js';
 export default class WorkspacesList extends Base {
   static description = 'List your workspaces';
 
@@ -12,7 +11,7 @@ export default class WorkspacesList extends Base {
   static enableJsonFlag = true;
 
   async run(): Promise<any> {
-    const xata = await getXataClient(this);
+    const xata = await this.getXataClient();
     const workspacesList = await xata.workspaces.getWorkspacesList();
 
     if (this.jsonEnabled()) return workspacesList.workspaces;
